+++ CHANGELOG.md
@@ -5,6 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Add `query_range()` to query an index for keys between A and B.
+- Add `base32` module with base32hex implementation to support numeric keys.
+
 ## [2.0.0] - 2019-12-23
 ### Changed
 - Rename `new()` to `open_default()`, and `with_opts()` to `open()`, to better